all Stats

+++ Assets/Bow.cs
@@ -8,7 +8,7 @@ public class Bow : Weapon
     public override int DMG(Stats stats)
     {
         float modifiedBase = BasicDamage + (stats.Agility * 4);
-        float dmg = modifiedBase * ((stats.Attack + 100) / 100f);
+        float dmg = modifiedBase * ((float)(stats.Attack + 100) / 100f);
         return Mathf.RoundToInt(dmg);
     }
 }
